+++ pocketbase/pb_migrations/1766133280_updated_emotions.js
@@ -1,0 +1,53 @@
+/// <reference path="../pb_data/types.d.ts" />
+migrate((app) => {
+  const collection = app.findCollectionByNameOrId("pbc_2502012209")
+
+  // remove field
+  collection.fields.removeById("file3309110367")
+
+  // add field
+  collection.fields.addAt(2, new Field({
+    "cascadeDelete": false,
+    "collectionId": "pbc_3607937828",
+    "hidden": false,
+    "id": "relation3309110367",
+    "maxSelect": 1,
+    "minSelect": 0,
+    "name": "image",
+    "presentable": false,
+    "required": false,
+    "system": false,
+    "type": "relation"
+  }))
+
+  return app.save(collection)
+}, (app) => {
+  const collection = app.findCollectionByNameOrId("pbc_2502012209")
+
+  // add field
+  collection.fields.addAt(2, new Field({
+    "hidden": false,
+    "id": "file3309110367",
+    "maxSelect": 1,
+    "maxSize": 0,
+    "mimeTypes": [
+      "image/jpeg",
+      "image/png",
+      "image/svg+xml",
+      "image/gif",
+      "image/webp"
+    ],
+    "name": "image",
+    "presentable": true,
+    "protected": false,
+    "required": true,
+    "system": false,
+    "thumbs": [],
+    "type": "file"
+  }))
+
+  // remove field
+  collection.fields.removeById("relation3309110367")
+
+  return app.save(collection)
+})