+++ sveltekit/src/lib/pb/constants.ts
@@ -1,0 +1,14 @@
+export const COLLECTION_NAMES = {
+	Classrooms: 'classrooms',
+	Decks: 'decks',
+	Slides: 'slides',
+	Presentations: 'presentations',
+	Images: 'images',
+	Emotions: 'emotions',
+	EmotionOverrides: 'emotion_overrides',
+	// People Collections
+	People: 'people',
+	// Activities Collections
+	PresentationActivity: 'presentation_activity',
+	PeopleActivity: 'people_activity'
+} as const;