+++ sveltekit/src/lib/slides/title/_types.ts
@@ -1,5 +1,5 @@
-import { Slide, PresentationActivity, StudentActivity, BaseActivityObject } from '$lib/pb/objects';
-import type { ActivityBasePB } from '$lib/pb/schema';
+import { BaseActivityObject, PresentationActivity, Slide, StudentActivity } from '$lib/pb';
+import type { ActivityBasePB } from '$lib/pb/schema/index';
 import type {
 	SlideComponentProps,
 	StudentInteractComponentProps,